# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Datespark API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Allows to login to service",
                "operationId": "5e09c26a0abb2b1bf040badb2d9c69ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "$ref": "#/components/schemas/UserUpdateDto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns user with updated schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Invalid body"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Allows user to delete its account",
                "operationId": "f8d7aea516ee464d0afee90f3a1a1722",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return id of deleted user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/DeletedModelResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/onboarding": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Finish user registration",
                "operationId": "1118bf0c204175a660438529fe96ace7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserOnboardingInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return data of user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "409": {
                        "description": "User already finished onboarding"
                    },
                    "422": {
                        "description": "Invalid body"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserOnboardingInput": {
                "description": "Additional data of user",
                "properties": {
                    "bio": {
                        "type": "string",
                        "example": "Example bio"
                    },
                    "fbLink": {
                        "type": "string",
                        "example": "https://facebook.com"
                    },
                    "igLink": {
                        "type": "string",
                        "example": "https://instagram.com"
                    },
                    "preferredGender": {
                        "description": "Either MALE or FEMALE",
                        "type": "string",
                        "example": "MALE"
                    },
                    "avatar": {
                        "description": "Base64 encoded image",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserUpdateDto": {
                "description": "User update",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Jan"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Kowalski"
                    },
                    "email": {
                        "type": "string",
                        "example": "jan@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    },
                    "gender": {
                        "description": "Either MALE or FEMALE",
                        "type": "string",
                        "example": "MALE"
                    }
                },
                "type": "object"
            },
            "DeletedModelResource": {
                "description": "Delete user",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "description": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jan Kowalski"
                    },
                    "surname": {
                        "type": "string",
                        "example": "jan@example.com"
                    },
                    "email": {
                        "type": "string",
                        "example": "jan@example.com"
                    },
                    "status": {
                        "description": "Either ACTIVE or DURING_REGISTER",
                        "type": "string",
                        "example": "ACTIVE"
                    },
                    "gender": {
                        "description": "Either MALE or FEMALE",
                        "type": "string",
                        "example": "MALE"
                    },
                    "roles": {
                        "description": "Array of roles of user in system",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "USER",
                                "ADMIN"
                            ]
                        },
                        "example": [
                            "USER"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}